requirements.txt
python-dotenv
SQLAlchemy
SQLAlchemy-Utils
psycopg2